+++ lib/assembly/image/jp2_creator.rb
