+++ spec/assembly/image_spec.rb
